Fixed bugs
Redirecting from index to other page was buggy; corrected the redirect
link from index.html to other.html so it functioned as it was supposed
to. index2.html is replaced by other.html to easier differentiate in
code between index and the second page, but the content is the same.